+++ react/components/LoginComponent.js
@@ -58,6 +58,12 @@ class LoginComponent extends Component {
       runtime: { history, navigate },
     } = this.props
 
+    const {
+      isAuthenticated,
+      firstName,
+      email
+    } = sessionProfile || {}
+
     const pathname = history && history.location && history.location.pathname
     const search = history && history.location && history.location.search
 
@@ -66,12 +72,12 @@ class LoginComponent extends Component {
 
     const buttonContent = hideIconLabel ? (
       <></>
-    ) : sessionProfile ? (
+    ) : isAuthenticated ? (
       <span
         className={`${styles.profile} truncate t-action--small order-1 pl4 ${labelClasses} dn db-l`}
       >
         {translate('store/login.hello', intl)}{' '}
-        {sessionProfile.firstName || sessionProfile.email}
+        {firstName || email}
       </span>
     ) : (
       <span
@@ -82,7 +88,7 @@ class LoginComponent extends Component {
     )
 
     if (loginButtonAsLink) {
-      const linkTo = sessionProfile ? 'store.account' : 'store.login'
+      const linkTo = isAuthenticated ? 'store.account' : 'store.login'
       return (
         <div className={styles.buttonLink}>
           <ButtonWithIcon
@@ -98,7 +104,7 @@ class LoginComponent extends Component {
                 return
               }
               const returnUrl =
-                !sessionProfile &&
+                !isAuthenticated &&
                 encodeURIComponent(`${pathname}${search}`)
               const bindingAddress = getBindingAddress()
               return navigate({
@@ -153,6 +159,8 @@ class LoginComponent extends Component {
      googleOneTapMarginTop,
     } = this.props
 
+    const { isAuthenticated } = sessionProfile || {}
+
     return (
       <div className={`${styles.container} flex items-center fr`}>
         <div className="relative">
@@ -161,7 +169,7 @@ class LoginComponent extends Component {
             <OneTapSignin
               alignment={googleOneTapAlignment}
               marginTop={googleOneTapMarginTop}
-              shouldOpen={!sessionProfile}
+              shouldOpen={!isAuthenticated}
             />
           )}
           {isBoxOpen && (